requirements-demo.txt
streamlit>=1.43,<2
google-api-python-client>=2.169,<3
google-auth-httplib2>=0.2.0,<1
google-auth-oauthlib>=1.2.1,<2
